backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
python-multipart
python-dotenv
requests
torch
torchvision
transformers
pillow
networkx
